Add files via upload
This project focuses on detecting influential entities on the Ethereum blockchain. In
particular, it models USD Tether Ethereum ERC-20 token transactions as a complex
network and uses degree, closeness, betweenness and PageRank centralities to identify the most influential addresses in the network. Furthermore, the study combines
centrality algorithms with the Louvain community detection algorithm to determine
the extent to which the most influential addresses tend to cluster with each other.
It also uses the Louvain algorithm in conjunction with the centrality metrics to assess how the relative importance of the most influential addresses changes on the
community level. The experimental results show that the USDT transaction network
exhibits a strong centralisation tendency. However, a significant clustering tendency
among the influential entities was not identified. Furthermore, it was found that the
relative importance of addresses with high PageRank and Betweenness centrality
tends to reduce on the community level.